add transport trait
